+++ src/components/datetime/DateTimeController.cpp
@@ -26,7 +26,7 @@ void DateTime::SetCurrentTime(std::chrono::time_point<std::chrono::system_clock,
   xSemaphoreGive(mutex);
 }
 
-void DateTime::SetTime(uint16_t year, uint8_t month, uint8_t day, uint8_t hour, uint8_t minute, uint8_t second) {
+void DateTime::SetTime(uint16_t year, uint8_t month, uint16_t day, uint8_t hour, uint8_t minute, uint8_t second) {
   std::tm tm = {
     /* .tm_sec  = */ second,
     /* .tm_min  = */ minute,
